+++ R/rdbnomics.R
@@ -10,6 +10,9 @@
 NULL
 
 if (getRversion() >= "2.15.1") {
-  vars <- c(".", ":=", "value", "dotI", "period", "period_start_day")
+  vars <- c(
+    ".", ":=", "value", "dotI", "period", "period_start_day", "series_code",
+    "filtered", "original_period", "series_name"
+  )
   utils::globalVariables(unique(vars))
 }