+++ NEWS.md
@@ -1,3 +1,10 @@
+# rdbnomics 0.5.0
+
+* New filters tool from <https://editor.nomics.world/filters>.
+* If the retrieved dataset contains columns with codes (like ISO codes), then
+  correspondences are performed to translate these codes if possible.
+* Internal function `get_data` better handles errors. The message is clearer.
+
 # rdbnomics 0.4.7
 
 * Use of `metadata` from the API to download less data.